# specs/004-task-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Task Management API
  description: |
    REST API for task management in the Todo Web Application.

    ## Authentication
    All endpoints require JWT token in Authorization header (Module 4).
    For development, uses hardcoded test user.

    ## Rate Limiting
    100 requests per minute per user. Returns 429 when exceeded.
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Development

tags:
  - name: Tasks
    description: Task CRUD operations

paths:
  /api/{user_id}/tasks:
    get:
      operationId: listTasks
      summary: List user's tasks
      description: Returns paginated list of tasks with optional filtering and sorting.
      tags: [Tasks]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Filter by completion status
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [created, title]
            default: created
        - name: cursor
          in: query
          description: Pagination cursor from previous response
          schema:
            type: string
        - name: limit
          in: query
          description: Number of tasks per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'

    post:
      operationId: createTask
      summary: Create a new task
      tags: [Tasks]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              minimal:
                summary: Title only
                value:
                  title: "Buy groceries"
              full:
                summary: With description
                value:
                  title: "Read book"
                  description: "Finish chapter 5"
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/{user_id}/tasks/{task_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      - name: task_id
        in: path
        required: true
        schema:
          type: integer

    get:
      operationId: getTask
      summary: Get a single task
      tags: [Tasks]
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateTask
      summary: Update a task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteTask
      summary: Delete a task
      tags: [Tasks]
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDeleteResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      operationId: toggleTaskComplete
      summary: Toggle task completion status
      tags: [Tasks]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional description

    TaskUpdate:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
          nullable: true
        completed:
          type: boolean

    TaskResponse:
      type: object
      required: [id, user_id, title, completed, created_at, updated_at]
      properties:
        id:
          type: integer
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    TaskListResponse:
      type: object
      required: [tasks, has_more]
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        next_cursor:
          type: string
          nullable: true
          description: Cursor for next page
        has_more:
          type: boolean
          description: Whether more tasks exist

    TaskDeleteResponse:
      type: object
      required: [message, task_id]
      properties:
        message:
          type: string
          enum: [deleted]
        task_id:
          type: integer

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, FORBIDDEN, UNAUTHORIZED, RATE_LIMITED]
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Invalid request
            details:
              - field: title
                message: String should have at least 1 character

    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Authentication required

    Forbidden:
      description: User ID mismatch
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Cannot access another user's tasks

    NotFound:
      description: Task not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Task not found

    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RATE_LIMITED
            message: Too many requests. Try again in 60 seconds.
      headers:
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer